worked on the linter errors
