+++ src/index.js
@@ -1,11 +1,4 @@
-// import pageLoad from "./page_load";
-// import aboutLoad from "./about";
-// import galeryLoad from "./galery";
-// import contactLoad from "./contact";
-import { tabsLoad, Tabs } from "./testtas";
-// import { navCnt } from "./nav";
-// import { galeryLoad } from "./galery";
-// import { mainnavCnt } from "./mainnav";
+import { Tabs } from './tabs';
 // create tabs
-const tabs = new Tabs(document.querySelector(".tabs"));
+const tabs = new Tabs(document.querySelector('.tabs'));
 tabs.init();
